pyproject.toml
[project]
name = "market-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "firecrawl-py>=4.12.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "langchain-groq>=1.1.1",
    "tavily-python>=0.7.17",
]
